pyproject.toml
[project]
name = "fhra"
version = "0.1.0"
description = "Family History Research Assistant — FamilySearch API tooling and research agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "ged4py>=0.5",
    "mcp>=1.0",
    "typer>=0.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
fhra = "fhra.cli:app"

[project.license]
file = "LICENSE"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fhra"]

# Ensure non-Python assets (schema, etc.) are included in built wheels.
[tool.hatch.build.targets.wheel.force-include]
"src/fhra/db/schema.sql" = "fhra/db/schema.sql"

[tool.hatch.build.targets.sdist]
include = [
    "src/fhra",
    "tests",
    "pyproject.toml",
    "README.md",
    "LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
